fix: create a common connectDB function
