+++ src/app/book/api/fetchBooks/route.ts
@@ -1,5 +1,4 @@
-import { open } from 'sqlite'
-import { verbose } from "sqlite3";
+import { connectDB } from "../utils";
 export interface Book {
     id: string,
     title: string,
@@ -16,22 +15,6 @@ export interface Book {
     img_url: string
 }
 
-const sqlite3 = verbose();
-
-async function connectDB() {
-    let DB = null;
-    try {
-        DB = await open({
-            filename: 'src/db/public_10book.db',
-            driver: sqlite3.Database
-        });
-        console.log('DB connection is ready');
-    }
-    catch (err) {
-        console.log(err)
-    }
-    return DB;
-}
 
 export async function GET() {
     const db = await connectDB();
