+++ src/app/book/api/fetchChaptersByBookId/route.ts
@@ -1,6 +1,4 @@
-import { open } from 'sqlite'
-import { verbose } from "sqlite3";
-const sqlite3 = verbose();
+import { connectDB } from "../utils";
 
 export interface Chapter {
     id: string,
@@ -9,21 +7,6 @@ export interface Chapter {
     chapter_name: string,
     content: string,
     nft_ip_id: string
-}
-
-async function connectDB() {
-    let DB = null;
-    try {
-        DB = await open({
-            filename: 'src/db/public_10book.db',
-            driver: sqlite3.Database
-        });
-        console.log('DB connection is ready');
-    }
-    catch (err) {
-        console.log(err)
-    }
-    return DB;
 }
 
 export async function GET(request: Request) {
